+++ operator/client/clientset/versioned/scheme/register.go
@@ -6,7 +6,7 @@
 package scheme
 
 import (
-	verrazzanov1alpha1 "github.com/verrazzano/verrazzano/operator/apis/verrazzano/v1alpha1"
+	verrazzanov1alpha1 "github.com/verrazzano/verrazzano/operator/api/verrazzano/v1alpha1"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	runtime "k8s.io/apimachinery/pkg/runtime"
 	schema "k8s.io/apimachinery/pkg/runtime/schema"
