+++ operator/client/clientset/versioned/typed/verrazzano/v1alpha1/fake/fake_verrazzano.go
@@ -8,7 +8,7 @@ package fake
 import (
 	"context"
 
-	v1alpha1 "github.com/verrazzano/verrazzano/operator/apis/verrazzano/v1alpha1"
+	v1alpha1 "github.com/verrazzano/verrazzano/operator/api/verrazzano/v1alpha1"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	labels "k8s.io/apimachinery/pkg/labels"
 	schema "k8s.io/apimachinery/pkg/runtime/schema"
