+++ operator/client/clientset/versioned/typed/verrazzano/v1alpha1/verrazzano.go
@@ -9,7 +9,7 @@ import (
 	"context"
 	"time"
 
-	v1alpha1 "github.com/verrazzano/verrazzano/operator/apis/verrazzano/v1alpha1"
+	v1alpha1 "github.com/verrazzano/verrazzano/operator/api/verrazzano/v1alpha1"
 	scheme "github.com/verrazzano/verrazzano/operator/client/clientset/versioned/scheme"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	types "k8s.io/apimachinery/pkg/types"
