+++ operator/controllers/upgrade_reconciler.go
@@ -6,7 +6,7 @@ package controllers
 import (
 	"fmt"
 
-	installv1alpha1 "github.com/verrazzano/verrazzano/operator/apis/verrazzano/v1alpha1"
+	installv1alpha1 "github.com/verrazzano/verrazzano/operator/api/verrazzano/v1alpha1"
 	"github.com/verrazzano/verrazzano/operator/internal/component"
 	"go.uber.org/zap"
 	ctrl "sigs.k8s.io/controller-runtime"
